Version 0.6.1
unitefsns(): bugfix
asaxspureresonant(): possibility to return unscaled separated curves
basicfittinggui(): a new version was created while retaining the older one as basicfittinggui_old(). The new one is called by default, and if the dataset is NOT a SASDict, the older one is called.
SASDict: substantial work was done on plotting and fitting. Concept of transformations is introduced. Online documentation is needed!